requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.6
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docx2txt==0.9
ecdsa==0.19.1
email-validator==2.3.0
emoji==2.14.1
et_xmlfile==2.0.0
fastapi==0.116.1
ffmpeg-python==0.2.0
filetype==1.2.0
frozenlist==1.7.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html2text==2025.4.15
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.28
langchain-google-genai==2.0.11
langchain-text-splitters==0.3.9
langdetect==1.0.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
lxml==6.0.1
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
nltk==3.9.1
numpy==2.3.2
olefile==0.47
openai==1.99.3
openpyxl==3.1.5
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
passlib==1.7.4
Pillow==11.0.0
playwright==1.54.0
postgrest==2.19.0
propcache==0.3.2
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
pypdf==5.9.0
python-docx==1.1.2
fpdf2==2.8.2
python-dotenv==1.1.1
python-iso639==2025.2.18
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.14.1
realtime==2.19.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
storage3==2.19.0
StrEnum==0.4.15
supabase==2.19.0
supabase-auth==2.19.0
supabase-functions==2.19.0
tenacity==9.1.2
tiktoken==0.10.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
unstructured==0.18.14
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
